requirement.txt
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
mysqlclient==2.1.1
